bug fix and add test for retries
