+++ sdk/core/System.ClientModel/src/Pipeline/PipelineTransport.cs
@@ -22,7 +22,7 @@ public abstract class PipelineTransport : PipelinePolicy
             throw new InvalidOperationException("Response was not set by transport.");
         }
 
-        message.Response.SetIsError(message.MessageClassifier?.IsErrorResponse(message) ?? default);
+        message.Response.SetIsError(ClassifyResponse(message));
     }
 
     /// <summary>
@@ -38,8 +38,12 @@ public abstract class PipelineTransport : PipelinePolicy
             throw new InvalidOperationException("Response was not set by transport.");
         }
 
-        message.Response.SetIsError(message.MessageClassifier?.IsErrorResponse(message) ?? default);
+        message.Response.SetIsError(ClassifyResponse(message));
     }
+
+    private static bool ClassifyResponse(PipelineMessage message) =>
+        message.MessageClassifier?.IsErrorResponse(message) ??
+        PipelineMessageClassifier.Default.IsErrorResponse(message);
 
     protected abstract void ProcessCore(PipelineMessage message);
